% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{Hello, world!}
\usage{
hello(x)
}
\arguments{
\item{x}{the name of the person saying hello.}
}
\value{
the output from \code{\link{print}}
}
\description{
Hello, world!
}
\examples{
hello("John")
\dontrun{
hello("James")
}
}
