% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{pss}
\alias{pss}
\title{Generate a request to BoF-PSS 3 object}
\usage{
pss(..., method = "GET", pss_domain = "http://localhost:8080")
}
\arguments{
\item{...}{query parameters}

\item{method}{request method}

\item{pss_domain}{domain of BoF-PSS 3 API, default https://localhost:8080}
}
\value{
a request to BoF-PSS 3 object
}
\description{
Generate a request to BoF-PSS 3 object
}
