% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-projects.R
\name{pss_create_proj}
\alias{pss_create_proj}
\title{Create a new BoF-PSS3 project}
\usage{
pss_create_proj(proj_name)
}
\arguments{
\item{proj_name}{project name}
}
\value{
a response with HTTP status code
}
\description{
Create a new BoF-PSS3 project
}
