% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{pss_delete}
\alias{pss_delete}
\title{Send a DELETE request to BoF-PSS 3}
\usage{
pss_delete(
  ...,
  data = NULL,
  path = NULL,
  type = "json",
  resp_type = "json",
  pss_domain = "http://localhost:8080"
)
}
\arguments{
\item{...}{query parameters}

\item{pss_domain}{domain of BoF-PSS 3 API, default https://localhost:8080}

\item{body}{JSON body for the request}
}
\value{
a response from BoF-PSS
}
\description{
Send a DELETE request to BoF-PSS 3
}
