% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-projects.R
\name{pss_delete_proj}
\alias{pss_delete_proj}
\title{Delete a BoF-PSS 3project}
\usage{
pss_delete_proj(proj_name)
}
\arguments{
\item{proj_name}{project name}
}
\value{
a response with HTTP status code
}
\description{
Delete a BoF-PSS 3project
}
