% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{pss_dry_run}
\alias{pss_dry_run}
\title{Perform a dry run on the request on BoF-PSS 3}
\usage{
pss_dry_run(
  ...,
  body = NULL,
  method = "GET",
  pss_domain = "http://localhost:8080"
)
}
\arguments{
\item{...}{query parameters}

\item{body}{JSON body for the request}

\item{method}{request method}

\item{pss_domain}{domain of BoF-PSS 3 API, default https://localhost:8080}
}
\value{
a list containing information on the request to BoF-PSS 3 object
}
\description{
Perform a dry run on the request on BoF-PSS 3
}
