% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{pss_get}
\alias{pss_get}
\title{Send a GET request to BoF-PSS 3}
\usage{
pss_get(..., resp_type = "json", pss_domain = "http://localhost:8080")
}
\arguments{
\item{...}{query parameters}

\item{pss_domain}{domain of BoF-PSS 3 API, default https://localhost:8080}
}
\value{
a response from BoF-PSS
}
\description{
Send a GET request to BoF-PSS 3
}
