% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{pss_info}
\alias{pss_info}
\title{Get BoF-PSS3 Simulator Information}
\usage{
pss_info(pss_domain = "http://localhost:8080")
}
\arguments{
\item{pss_domain}{domain of BoF-PSS 3 API, default https://localhost:8080}
}
\value{
a list containing information about installed BoF-PSS 3 application
}
\description{
Get BoF-PSS3 Simulator Information
}
