% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-projects.R
\name{pss_search_proj}
\alias{pss_search_proj}
\title{List existing BoF-PSS 3 project(s)}
\usage{
pss_search_proj(proj_name = NULL)
}
\arguments{
\item{proj_name}{project name}
}
\value{
list of projects
}
\description{
List existing BoF-PSS 3 project(s)
}
