% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{pss_upload}
\alias{pss_upload}
\title{Send a request to upload files to BoF-PSS 3}
\usage{
pss_upload(..., path, pss_domain = "http://localhost:8080")
}
\arguments{
\item{...}{query parameters}

\item{pss_domain}{domain of BoF-PSS 3 API, default https://localhost:8080}

\item{file}{file to upload}
}
\value{
a response from BoF-PSS
}
\description{
Send a request to upload files to BoF-PSS 3
}
